app/requirements.txt
db-sqlite3
flask-jwt-extended
requests
gunicorn